+++ docs/reference/changelog.md
@@ -7,6 +7,12 @@ categories: reference
 
 #### 0.54 (unreleased)
 
+Features:
+
+* `--method vpn-tcp` now works on minikube and minishift.
+  As a result we now recommend using it as the default method.
+  ([#160](https://github.com/datawire/telepresence/issues/160))
+
 Bug fixes:
 
 * Support more versions of Linux in container mode.